myblog/requirements.txt
Django<=1.7
psycopg2
PIL
-e git://github.com/kennethreitz/dj-static.git#egg=dj-static
dj-database-url>=0.2.2
django-autoslug>=1.7.1
gunicorn>=18.0
South>=0.8.4
model-mommy>=1.2